+++ src/Facades/Visits.php
@@ -9,6 +9,7 @@ use Illuminate\Support\Facades\Facade;
 
 /**
  * @method static void track(string $name, ?\Illuminate\Database\Eloquent\Model $eventable = null, ?array $meta = null, ?string $inheritFrom = null)
+ * @method static void identify(\Illuminate\Contracts\Auth\Authenticatable $user)
  * @method static array whoami(?\Illuminate\Http\Request $request = null, ?string $ip = null)
  */
 class Visits extends Facade